Adjust import formatting for isort 5.6.x

+++ av/container/pyio.pxd
@@ -1,4 +1,4 @@
-from libc.stdint cimport uint8_t, int64_t
+from libc.stdint cimport int64_t, uint8_t
 
 
 cdef int pyio_read(void *opaque, uint8_t *buf, int buf_size) nogil
